% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andersen.R
\name{andersenplot}
\alias{andersenplot}
\title{Plot hazards for two strata for each time.  At times with an event in one but not the other group, the fitted hazard remains constant, and so the plot is a step function.  If hazards are proportional between strata, then the plot should be close to a straight line.}
\usage{
andersenplot(fit)
}
\arguments{
\item{fit}{A coxph fit with a stratification term.}
}
\description{
Plot hazards for two strata for each time.  At times with an event in one but not the other group, the fitted hazard remains constant, and so the plot is a step function.  If hazards are proportional between strata, then the plot should be close to a straight line.
}
