% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arjas.R
\name{arjasplot}
\alias{arjasplot}
\title{Examine the potential role of treatment in treatment in 
a model already including sex.
Straight lines that are not 45 degrees
indicate the appropriateness of new variable as a linear
effect.}
\usage{
arjasplot(formulastring, time, stratifier, status, mydata)
}
\arguments{
\item{formulastring}{A formula for a coxph fit.}

\item{time}{The name of the time variable}

\item{stratifier}{The name of the stratifier variable}

\item{status}{The name of the status variable}

\item{mydata}{The data frame.}
}
\description{
Examine the potential role of treatment in treatment in 
a model already including sex.
Straight lines that are not 45 degrees
indicate the appropriateness of new variable as a linear
effect.
}
