% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkcensor.R
\name{checkcensor}
\alias{checkcensor}
\title{Check how censoring impacts sampling properties of KM fit and log rank test.}
\usage{
checkcensor(nsamp = 1000, nobs = 1000)
}
\arguments{
\item{nsamp}{Number of MC samples}

\item{nobs}{Number of observations}
}
\value{
biases of fits.
}
\description{
Check how censoring impacts sampling properties of KM fit and log rank test.
}
