% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkresults.R
\name{checkresults}
\alias{checkresults}
\title{Produce a graphical assessment of Monte Carlo experiment on fidelity of proportional hazards regression to the uniform ideal.}
\usage{
checkresults(regnsimulation, frac = 0.1)
}
\arguments{
\item{regnsimulation}{A structure with a component out, matrix with columns representing definitions of p-values and as many rows as there MC samples.}

\item{frac}{Proportion for bottom of distribution to be assessed.}
}
\value{
A list with components of consisting of simulated Wald p-values, likelihood ratio p-values, and corrected likelihood ratio p-values.
}
\description{
This function draws a quantile plot for Monte Carlo assessments of fit to the corrected proportional hazards fit.
}
