% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareplot.R
\name{compareplot}
\alias{compareplot}
\title{Plot resuts of simcode}
\usage{
compareplot(simresults)
}
\arguments{
\item{simresults}{the result of simcode}
}
\value{
nothing.
}
\description{
Plot resuts of simcode
}
