% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compete.simulation.R
\name{compete.simulation}
\alias{compete.simulation}
\title{Simulate from a competing risk model with correlated log normal errors, and plot various estimates.}
\usage{
compete.simulation(ncr = 4, sig = 0.8, ns = 1000)
}
\arguments{
\item{ncr}{Number of competing risks.}

\item{sig}{correlation among competing risks.}

\item{ns}{number of observations.}
}
\description{
Simulate from a competing risk model with correlated log normal errors, and plot various estimates.
}
