% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertmtol.R
\name{convertmtol}
\alias{convertmtol}
\title{Convert a polytomous regression to a conditional logistic regression.}
\usage{
convertmtol(xmat, str, yvec, subjects)
}
\arguments{
\item{xmat}{regression matrix}

\item{str}{stratum label}

\item{yvec}{vector of responses}

\item{subjects}{vector of subject labels passed directly to the output.}
}
\value{
a data set on which to apply conditional logistic regression, corresponding to the multinomial regression model.
}
\description{
Convert a polytomous regression to a conditional logistic regression.
}
\details{
Implements version of \insertCite{kolassa16}{PHInfiniteEstimates}.  
The multinomial regression is converted to a conditional logistic regression, and methods of \insertCite{kolassa97}{PHInfiniteEstimates} may be applied.
This function differs from \code{convertbaselineltolr} of this package in that the former treats the richer data structure of package \code{mlogit}, and this function treats a less complicated structure.
Data in the example is the breast cancer data set \code{breast} of package \code{coxphf}.
}
\examples{
#Uses data set breast from package coxphf.
out<-convertstoml(Surv(breast$TIME,breast$CENS),breast[,c("T","N","G","CD")])
out1<-convertmtol(out[,c("T","N","G","CD")],out[,"chid"],out[,"choice"],
   out[,"patients"])
glmout<-glm.fit(out1$xmat,out1$y,family=binomial())
#In many practice examples, the following line shows which observations to retain
#in the logistic regression example.
moderate<-(fitted(glmout)<1-1.0e-8)&(fitted(glmout)>1.0e-8)
# Proportional hazards fit illustrating infinite estimates.
coxph(Surv(TIME,CENS)~ T+ N+ G+ CD,data=breast)
# Wrong analysis naively removing covariate with infinite estimate
coxph(Surv(TIME,CENS)~ T+ N+ CD,data=breast)
summary(glm((CENS>22)~T+N+G+CD,family=binomial,data=breast))
\donttest{
out2<-reduceLR(out1$xmat,yvec=out1$y,keep="CD")
bestcoxout<-coxph(Surv(TIME,CENS)~ T+ N+ G+ CD,data=breast,
   subset=as.numeric(unique(out1$subjects[out2$moderate])))
}
}
\references{
\insertRef{kolassa97}{PHInfiniteEstimates}

\insertRef{kolassa16}{PHInfiniteEstimates}
}
