% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gehan.wilcoxon.R
\name{gehan.wilcoxon.test}
\alias{gehan.wilcoxon.test}
\title{Perform Gehan's application to the Wilcoxon test for multiple samples, testing for equivalance of survival curve.  See Klein and Moeschberger (1997) Survival Analysis (7.3.3) and pp. 193-194.}
\usage{
gehan.wilcoxon.test(myformula, data, gehan = TRUE)
}
\arguments{
\item{myformula}{Proportional hazards formula appropriate for survfit}

\item{data}{the data set}

\item{gehan}{logical flag triggering the Wilcoxon test (gehan=TRUE), with weights equal to total at risk, or the log rank test (gehan=FALSE) with weights all 1.}
}
\value{
An htest-like object with the chi-square version of the test.
}
\description{
Perform Gehan's application to the Wilcoxon test for multiple samples, testing for equivalance of survival curve.  See Klein and Moeschberger (1997) Survival Analysis (7.3.3) and pp. 193-194.
}
