% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrapproximations.R
\name{lrapproximations}
\alias{lrapproximations}
\title{Assess the accuracy of the log rank statistic approximation to the true value, in the case without censoring.  Provides plots of statistics, and empirical test level.}
\usage{
lrapproximations(nobs = 10, ratio = 1, nsamp = 1000)
}
\arguments{
\item{nobs}{number of observations in each group.  This currently supports only equal group size data sets.}

\item{ratio}{Ratio of group means; use 1 for null.}

\item{nsamp}{Monte Carlo sample size.}
}
\value{
a vector of empirical test sizes.
}
\description{
Assess the accuracy of the log rank statistic approximation to the true value, in the case without censoring.  Provides plots of statistics, and empirical test level.
}
\examples{
lrapproximations(nsamp=100)
}
