% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simultaneouscoverage.R
\name{simultaneouscoverage}
\alias{simultaneouscoverage}
\title{Calculate simultaneous coverage of pointwise confidence intervals.}
\usage{
simultaneouscoverage(nsamp, nobs)
}
\arguments{
\item{nsamp}{Number of Monte Carlo samples.}

\item{nobs}{Number of observations per sample.}
}
\value{
Simultaneous coverage proportion.
}
\description{
Simulate exponential event times with expecation 1.  Simulate censoring times with expectation 2.  Calculate confidence intervals and check simultaneous coverage.
}
\examples{
simultaneouscoverage(1000,20)
}
