% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizefits.R
\name{summarizefits}
\alias{summarizefits}
\title{Summarize proportional hazards fits}
\usage{
summarizefits(
  repairedfit,
  penalizedout,
  penalizedoutsmaller,
  iv,
  verbose = TRUE
)
}
\arguments{
\item{repairedfit}{coxph fit}

\item{penalizedout}{coxphf fit}

\item{penalizedoutsmaller}{smaller coxphf fit}

\item{iv}{name of the variable of interest, as a character string}

\item{verbose}{logical flag triggering intermediate messaging.}
}
\value{
a vector with components
\itemize{
    \item Wald p-value from the Cox regression fit.
    \item partial likelihood ratio p-value from Cox regression fit.
    \item parameter estimate from the Cox regression fit.
    \item standard error from the Cox regression fit.
    \item Conditional Skovgaard standard error from the Cox regression fit.
    \item Signed root of the partial likelihood ratio statistic from Cox regression fit.
    \item partial likelihood ratio statistic p-value from coxphf
    \item Wald p-value from coxphf
    \item parameter estimate from coxphf
    \item standard error from coxphf
    \item number of parameters in reduced fit.
}
}
\description{
Summarize proportional hazards fits
}
\references{
\insertRef{kz19}{PHInfiniteEstimates}
}
