% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizetable.R
\name{summarizetable}
\alias{summarizetable}
\title{Summarize the results of simulations investigating operating conditions for the data reduction method to avoid monotone likelihood.  Files are of form "hsxxx", for xxx numerals.}
\usage{
summarizetable()
}
\description{
Summarize the results of simulations investigating operating conditions for the data reduction method to avoid monotone likelihood.  Files are of form "hsxxx", for xxx numerals.
}
