% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testcox.R
\name{testcox}
\alias{testcox}
\title{Test size of asymptotic Cox tests.}
\usage{
testcox(nsamp = 1000, nobs = 50, ncov = 5, randomcov = TRUE)
}
\arguments{
\item{nsamp}{Number of MC samples}

\item{nobs}{Number of observations}

\item{ncov}{Number of covariates}

\item{randomcov}{Indicator of whether to draw random covariates.}
}
\value{
level of two-sided test of nominal size 0.05.
}
\description{
Test size of asymptotic Cox tests.
}
