% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voter.ml.R
\docType{data}
\name{voter.ml}
\alias{voter.ml}
\title{Subset of British elections data used in \insertCite{kolassa16}{PHInfiniteEstimates}.  
Data are from \insertCite{bes}{PHInfiniteEstimates}.}
\usage{
data(voter.ml)
}
\description{
Subset of British elections data used in \insertCite{kolassa16}{PHInfiniteEstimates}.  
Data are from \insertCite{bes}{PHInfiniteEstimates}.
}
\examples{
data(voter.ml)
}
\references{
\insertRef{bes}{PHInfiniteEstimates}

\insertRef{kolassa16}{PHInfiniteEstimates}
}
