% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkresults.R
\name{checkresults}
\alias{checkresults}
\title{Produce a graphical assessment of Monte Carlo experiment on fidelity of proportional hazards regression to the uniform ideal.}
\usage{
checkresults(regnsimulation)
}
\arguments{
\item{regnsimulation}{A matrix with six columns and as many rows as there MC samples.}
}
\value{
A list with components of consisting of simulated Wald p-values, likelihood ratio p-values, and corrected likelihood ratio p-values.
}
\description{
This function draws a quantile plot for Monte Carlo assessments of fit to the corrected proportional hazards fit.
}
